+++ docs/changes.rst
@@ -37,6 +37,13 @@ Bug fixes
   the resulting output could be viewed by only some tools (i.e. github
   messed it up). Changed to display as a .rst file output. (see issue #27gi)
 
+* Fix issues around password so that we get the password from a prompt rather
+  than just from the command line. This also adds a form of persistence to
+  save info on connections by using the connection group.  It adds several
+  subcommands to the connection group to show, list, create, delete, etc.
+  connections so the user can maintain a list of WBEM servers that can be easily
+  referenced rather than having to type them in.(issue # 6))
+
 
 Build, test, quality
 ^^^^^^^^^^^^^^^^^^^^